Rename decorators to field
