+++ polygraph/types/tests/test_interface.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from polygraph.exceptions import PolygraphSchemaError
-from polygraph.types.decorators import field
+from polygraph.types.field import field
 from polygraph.types.interface import Interface, validate_interface_schema
 from polygraph.types.scalar import String
